backend/requirements.txt
asgiref==3.5.2
Django==4.0.4
djangorestframework==3.13.1
flake8==4.0.1
mccabe==0.6.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2022.1
sqlparse==0.4.2
